+++ Changes.md
@@ -1,3 +1,5 @@
+* #1071 - allow the user to delete a playoff bracket
+* #1071 - prevent the user from uninitializing a playoff bracket that will cause problems with run numbers
 * #1072 - Fix bug creating team finalist schedule
 * #1069 - support viewing a bye run
 * sort teams by table during head to head